+++ tests/test_action.py
@@ -24,6 +24,7 @@ def test_poweroff_trigger(delay):
                 homeassistant_discovery_prefix="homeassistant",
                 homeassistant_discovery_object_id="node",
                 poweroff_delay=delay,
+                monitored_system_unit_names=[],
             )
         )
     schedule_shutdown_mock.assert_called_once_with(action="poweroff", delay=delay)
@@ -45,6 +46,7 @@ def test_mqtt_topic_suffix_action_mapping_poweroff(topic_suffix, expected_action
                 homeassistant_discovery_prefix="homeassistant",
                 homeassistant_discovery_object_id="node",
                 poweroff_delay=datetime.timedelta(),
+                monitored_system_unit_names=[],
             )
         )
     login_manager_mock.ScheduleShutdown.assert_called_once()