+++ tests/test_mqtt.py
@@ -66,6 +66,7 @@ async def test__run(
             homeassistant_discovery_prefix=homeassistant_discovery_prefix,
             homeassistant_discovery_object_id=homeassistant_discovery_object_id,
             poweroff_delay=datetime.timedelta(),
+            monitored_system_unit_names=[],
         )
     assert caplog.records[0].levelno == logging.INFO
     assert caplog.records[0].message == (
@@ -169,6 +170,7 @@ async def test__run_tls(caplog, mqtt_host, mqtt_port, mqtt_disable_tls):
             homeassistant_discovery_prefix="homeassistant",
             homeassistant_discovery_object_id="host",
             poweroff_delay=datetime.timedelta(),
+            monitored_system_unit_names=[],
         )
     mqtt_client_class_mock.assert_called_once()
     _, mqtt_client_init_kwargs = mqtt_client_class_mock.call_args
@@ -204,6 +206,7 @@ async def test__run_tls_default():
             homeassistant_discovery_prefix="homeassistant",
             homeassistant_discovery_object_id="host",
             poweroff_delay=datetime.timedelta(),
+            monitored_system_unit_names=[],
         )
     mqtt_client_class_mock.assert_called_once()
     # enabled by default
@@ -236,6 +239,7 @@ async def test__run_authentication(
             homeassistant_discovery_prefix="discovery-prefix",
             homeassistant_discovery_object_id="node-id",
             poweroff_delay=datetime.timedelta(),
+            monitored_system_unit_names=[],
         )
     mqtt_client_class_mock.assert_called_once()
     _, mqtt_client_init_kwargs = mqtt_client_class_mock.call_args
@@ -267,6 +271,7 @@ async def test__run_authentication_missing_username(
                 homeassistant_discovery_prefix="discovery-prefix",
                 homeassistant_discovery_object_id="node-id",
                 poweroff_delay=datetime.timedelta(),
+                monitored_system_unit_names=[],
             )
     dbus_signal_loop_mock.assert_not_called()
 
@@ -295,6 +300,7 @@ async def test__run_sigint(mqtt_topic_prefix: str):
                 homeassistant_discovery_prefix="homeassistant",
                 homeassistant_discovery_object_id="host",
                 poweroff_delay=datetime.timedelta(),
+                monitored_system_unit_names=[],
             )
     async with mqtt_client_class_mock() as mqtt_client_mock:
         pass
@@ -327,6 +333,7 @@ async def test__mqtt_message_loop_trigger_poweroff(
         homeassistant_discovery_prefix="homeassistant",
         homeassistant_discovery_object_id="whatever",
         poweroff_delay=datetime.timedelta(seconds=21),
+        monitored_system_unit_names=[],
     )
     mqtt_client_mock = unittest.mock.AsyncMock()
     mqtt_client_mock.messages.__aiter__.return_value = [
@@ -374,6 +381,7 @@ async def test__mqtt_message_loop_retained(
         homeassistant_discovery_prefix="homeassistant",
         homeassistant_discovery_object_id="whatever",
         poweroff_delay=datetime.timedelta(seconds=21),
+        monitored_system_unit_names=[],
     )
     mqtt_client_mock = unittest.mock.AsyncMock()
     mqtt_client_mock.messages.__aiter__.return_value = [
@@ -402,3 +410,21 @@ async def test__mqtt_message_loop_retained(
             "ignoring retained message on topic 'systemctl/host/poweroff'",
         ),
     ]
+
+
+@pytest.mark.parametrize("mqtt_topic_prefix", ["systemctl/host", "systemd/raspberrypi"])
+@pytest.mark.parametrize("unit_name", ["foo.service", "bar.service"])
+def test_state_get_system_unit_active_state_mqtt_topic(
+    mqtt_topic_prefix: str, unit_name: str
+) -> None:
+    state = systemctl_mqtt._State(
+        mqtt_topic_prefix=mqtt_topic_prefix,
+        homeassistant_discovery_prefix="homeassistant",
+        homeassistant_discovery_object_id="whatever",
+        poweroff_delay=datetime.timedelta(seconds=21),
+        monitored_system_unit_names=[],
+    )
+    assert (
+        state.get_system_unit_active_state_mqtt_topic(unit_name=unit_name)
+        == f"{mqtt_topic_prefix}/unit/system/{unit_name}/active-state"
+    )