+++ CHANGELOG.md
@@ -10,11 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (https://github.com/fphammerle/systemctl-mqtt/issues/97)
 - birth & last will message on topic `systemctl/[hostname]/status`
   ("online" / "offline", https://github.com/fphammerle/systemctl-mqtt/issues/38)
+- command-line option `--monitor-system-unit [unit_name]` enables reports on
+  topic `systemctl/[hostname]/unit/system/[unit_name]/active-state`
+  (https://github.com/fphammerle/systemctl-mqtt/issues/56)
 - automatic discovery in home assistant:
   - availability status
   - entity `button.[hostname]_logind_lock_all_sessions`
   - entity `button.[hostname]_logind_poweroff`
   - entity `button.[hostname]_logind_suspend`
+  - entity `sensor.[hostname]_unit_system_[unit_name]_active_state`
+    for each command-line parameter `--monitor-system-unit [unit_name]`
 - command-line option `--log-level {debug,info,warning,error,critical}`
 - declare compatibility with `python3.11`, `python3.12` & `python3.13`